Remove keytar
