+++ package.json
@@ -590,7 +590,6 @@
     "@malloydata/syntax-highlight": "0.0.126",
     "@vscode/webview-ui-toolkit": "^1.2.1",
     "duckdb": "0.9.2",
-    "keytar": "7.7.0",
     "lit": "^3.0.2",
     "live-plugin-manager": "^0.18.1",
     "lodash": "^4.17.21",
