+++ src/extension/browser/commands/edit_connections.ts
@@ -22,7 +22,6 @@
  */
 
 import * as vscode from 'vscode';
-import {ConnectionConfig} from '../../../common/types/connection_manager_types';
 import {EditConnectionPanel} from '../../connection_editor';
 import {WorkerConnection} from '../../worker_connection';
 import {connectionConfigManager} from '../connection_config_manager_browser';
@@ -35,75 +34,8 @@ export function editConnectionsCommand(
   id?: string
 ): void {
   if (!panel) {
-    panel = new EditConnectionPanel(
-      connectionConfigManager,
-      worker,
-      (connections: ConnectionConfig[]) =>
-        handleConnectionsPreLoad(context, connections),
-      (connections: ConnectionConfig[]) =>
-        handleConnectionsPreSave(context, connections)
-    );
+    panel = new EditConnectionPanel(connectionConfigManager, context, worker);
     panel.onDidDispose(() => (panel = null));
   }
   panel.reveal(id);
-}
-
-/**
- * Perform setup on the connections being sent to the webview.
- *
- * @param connections The connection configs.
- * @returns An updated list of connections
- *
- * - Handles scrubbing passwords and putting them in the keychain.
- */
-// TODO(whscullin) keytar
-// This can be moved to common code once keytar is removed
-async function handleConnectionsPreLoad(
-  context: vscode.ExtensionContext,
-  connections: ConnectionConfig[]
-): Promise<ConnectionConfig[]> {
-  const modifiedConnections = [];
-  for (let connection of connections) {
-    if ('motherDuckToken' in connection && connection.motherDuckToken) {
-      const key = `connections.${connection.id}.motherDuckToken`;
-      const motherDuckToken = await context.secrets.get(key);
-      connection = {
-        ...connection,
-        motherDuckToken,
-      };
-    }
-    modifiedConnections.push(connection);
-  }
-  return modifiedConnections;
-}
-
-/**
- * Perform cleanup on the connections object received from the webview,
- * and prepare to save the config.
- *
- * @param connections The connection configs received from the webview.
- * @returns An updated list of connections
- *
- * - Handles scrubbing passwords and putting them in the keychain.
- */
-// TODO(whscullin) keytar
-// This can be moved to common code once keytar is removed
-async function handleConnectionsPreSave(
-  context: vscode.ExtensionContext,
-  connections: ConnectionConfig[]
-): Promise<ConnectionConfig[]> {
-  const modifiedConnections = [];
-  for (let connection of connections) {
-    if ('motherDuckToken' in connection && connection.motherDuckToken) {
-      const key = `connections.${connection.id}.motherDuckToken`;
-      await context.secrets.store(key, connection.motherDuckToken);
-      connection = {
-        ...connection,
-        // Change the config to trigger a connection reload
-        motherDuckToken: `$secret-${Date.now().toString()}$`,
-      };
-    }
-    modifiedConnections.push(connection);
-  }
-  return modifiedConnections;
 }
